Move tests outside app_helper package
Mostly cosmetic refactoring to ensure we have a cleaner app_helper package

Some tests required changes to ensure tests loading

- removed work_in context manager calls for migrations and runners tests
- added context manager to reset test environment in runners tests
