+++ changes/191.feature
@@ -1,0 +1,1 @@
+Reorganize tests outside main package
