+++ docs/conf.py
@@ -21,13 +21,12 @@ import sphinx_rtd_theme
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
-sys.path.insert(0, os.path.abspath("../app_helper/test_utils"))
-sys.path.insert(1, os.path.abspath(".."))
+sys.path.insert(0, os.path.abspath(".."))
 import helper  # isort:skip  # noqa
 import app_helper  # isort:skip  # noqa
 # fmt: on
 
-helper.setup_nocms()
+helper.setup()
 
 # -- General configuration ------------------------------------------------
