+++ requirements-docs.txt
@@ -1,3 +1,6 @@
 -r requirements-test.txt
 Django<3.0
 django-filer
+django-cms
+djangocms-admin-style
+djangocms-text-ckeditor
