+++ tests/test_utils/cms_helper_extra_runner.py
@@ -1,4 +1,4 @@
-from app_helper.test_utils.helper import HELPER_SETTINGS
+from helper import HELPER_SETTINGS
 
 HELPER_SETTINGS["TEMPLATE_CONTEXT_PROCESSORS"] = ["django.core.context_processors.debug"]
 HELPER_SETTINGS["TEST_RUNNER"] = "runners.CapturedOutputRunner"
