+++ tests/test_utils/example1/data/django_0001_partial.py
@@ -3,15 +3,15 @@ from django.db import migrations, models
 
 class Migration(migrations.Migration):
 
+    initial = True
+
     dependencies = []
 
     operations = [
         migrations.CreateModel(
             name="ExampleModel1",
             fields=[
-                ("id", models.AutoField(primary_key=True, serialize=False, verbose_name="ID", auto_created=True)),
+                ("id", models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name="ID")),
             ],
-            options={},
-            bases=(models.Model,),
         ),
     ]
