+++ tox.ini
@@ -14,7 +14,7 @@ envlist =
 skip_missing_interpreters = True
 
 [testenv]
-commands = {env:COMMAND:python} setup.py test {posargs}
+commands = {env:COMMAND:python} helper.py {posargs}
 deps=
     django22: django-mptt>0.9,<1.0
     django22: easy-thumbnails>2.4,<2.6
@@ -150,4 +150,6 @@ ignore =
     app_helper/test_utils/example1/locale/**
     changes/**
     docs/**
+    helper.py
     tasks.py
+    tests/**
